requirements.txt
celery>=4.2.0rc4
Django>=2.0.5
djangorestframework>=3.8.2
psycopg2>=2.7